Ticket 333 - Testsuitesmanagement versioncheck update
