+++ RobotFramework_TestsuitesManagement/Config/CConfig.py
@@ -35,6 +35,7 @@ from builtins import staticmethod
 
 import RobotFramework_TestsuitesManagement as TM
 from RobotFramework_TestsuitesManagement.Utils.CStruct import CStruct
+from RobotFramework_TestsuitesManagement.Utils.CVersion import bundle_version, INSTALLER_LOCATION, BUNDLE_NAME, BUNDLE_VERSION
 from PythonExtensionsCollection.String.CString import CString
 
 from JsonPreprocessor import CJsonPreprocessor
@@ -43,74 +44,7 @@ from robot.version import get_full_version, get_version
 from robot.libraries.BuiltIn import BuiltIn
 from robot.utils.dotdict import DotDict
 import pathlib
-from RobotFramework_TestsuitesManagement.version import VERSION, VERSION_DATE
 from RobotFramework_TestsuitesManagement.Utils.CVersion import CVersion, CVersionCheck
-
-INSTALLER_LOCATION = "https://github.com/test-fullautomation/robotframework-testsuitesmanagement/releases"
-BUNDLE_NAME = "RobotFramework_TestsuitesManagement"
-BUNDLE_VERSION = VERSION
-BUNDLE_VERSION_DATE = VERSION_DATE
-
-# Load package context file
-context_filename = "package_context.json"
-context_filepath = os.path.join(os.path.dirname(__file__), context_filename)
-context_config = None
-
-if os.path.isfile(context_filepath):
-    if os.stat(context_filepath).st_size == 0:
-        logger.warn(f"The '{context_filepath}' file is existing but empty.")
-    else:
-        package_context_schema = {
-            "type": "object",
-            'additionalProperties': False,
-            "properties": {
-                "installer_location": {"type": "string"},
-                "bundle_name": {"type": "string"},
-                "bundle_version": {"type": "string"},
-                "bundle_version_date": {"type": "string"}
-            }
-        }
-        try:
-            with open(context_filepath) as f:
-                context_config = json.load(f)
-        except Exception as reason:
-            logger.error(f"Cannot load the '{context_filepath}' file. Reason: {reason}")
-            exit(1)
-        
-        try:
-            validate(instance=context_config, schema=package_context_schema)
-        except Exception as reason:
-            logger.error(f"Invalid '{context_filepath}' file. Reason: {reason}")
-            exit(1)
-
-        if ('installer_location' in context_config) and context_config['installer_location']:
-            INSTALLER_LOCATION = context_config['installer_location']
-        if ('bundle_name' in context_config) and context_config['bundle_name']:
-            BUNDLE_NAME = context_config['bundle_name']
-        if ('bundle_version' in context_config) and context_config['bundle_version']:
-            BUNDLE_VERSION = context_config['bundle_version']
-        if ('bundle_version_date' in context_config) and context_config['bundle_version_date']:
-            BUNDLE_VERSION_DATE = context_config['bundle_version_date']
-
-def bundle_version():
-   '''
-This function prints out the package version which is:
-
-- RobotFramework_TestsuitesManagement version when this module is installed
-stand-alone (via `pip` or directly from sourcecode)
-
-- RobotFramework AIO version when this module is bundled with RobotFramework AIO
-package
-
-**Arguments:**
-
-* No input parameter is required
-
-**Returns:**
-
-* No return variable
-   '''
-   print(f"{BUNDLE_VERSION}")
 
 class CConfig():
     '''
@@ -685,7 +619,7 @@ In case the current version is not between min and max version, then the executi
 testsuite is terminated with "unknown" state
         '''
         oVersion = CVersion(self.sMinVersion, self.sMaxVersion)
-        res, reason = oVersion.verifyVersion(BUNDLE_VERSION)
+        res, reason = oVersion.verifyVersion()
         if res:
             if reason == CVersionCheck.WITHOUTVERSION.value:
                 logger.info(f"Running without {BUNDLE_NAME} version check!")
