+++ RobotFramework_TestsuitesManagement/Config/__init__.py
@@ -11,4 +11,4 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from .CConfig import CConfig, bundle_version, BUNDLE_VERSION
+from .CConfig import CConfig
