+++ RobotFramework_TestsuitesManagement/Utils/__init__.py
@@ -15,4 +15,4 @@
 
 from .LibListener import LibListener
 from .CStruct import CStruct
-from .CVersion import CVersion
+from .CVersion import CVersion, bundle_version, BUNDLE_VERSION, BUNDLE_NAME, INSTALLER_LOCATION
